商品出品機能　修正分
テスト修正しましたので、テスト結果の画像を以下に記載いたします。
https://gyazo.com/59e040bc19175e12e031282dcf15565a
