+++ app/controllers/application_controller.rb
@@ -12,7 +12,7 @@ class ApplicationController < ActionController::Base
   end
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:nickname, :email, :password, :encrypted_password, :first_name, :last_name, :first_name_yomi, :last_name_yomi, :birthday,])
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:nickname, :encrypted_password, :first_name, :last_name, :first_name_yomi, :last_name_yomi, :birthday,])
   end
 
 end
