+++ app/controllers/items_controller.rb
@@ -1,8 +1,8 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, except: [:index, :show]
+  before_action :authenticate_user!, except: [:index] #, :show]
 
   def index
-    @item = Item.includes(:user).order('created_at DESC')
+   # @items = Item.includes(:user).order('created_at DESC')
   end
 
   def new
@@ -18,9 +18,8 @@ class ItemsController < ApplicationController
     end
   end
 
-  def show
-
-  end
+  # def show
+  # end
 
 
   private
