+++ app/javascript/packs/application.js
@@ -3,22 +3,10 @@
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
 
-window.addEventListener('load', () => {
-  const priceInput = document.getElementById("item-price");
- //  console.log(priceInput)
-   priceInput.addEventListener("input", () => {
-     const addTaxDom = document.getElementById("add-tax-price");
-     // console.log(priceInput.innerHTML);
-     addTaxDom.innerHTML = Math.round(priceInput.value * 0.1 );
-     const addPofitDom = document.getElementById("profit");
-     // console.log(priceInput.innerHTML);
-     addPofitDom.innerHTML = Math.round(priceInput.value - Math.round(priceInput.value * 0.1 ))
- })
-});
-
 require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
+require("../price");
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
