+++ app/models/item.rb
@@ -1,22 +1,17 @@
 class Item < ApplicationRecord
   extend ActiveHash::Associations::ActiveRecordExtensions
 
-  # テーブルとのアソシエーション
   belongs_to :user
 
-  # アクティブハッシュとのアソシエーション
   belongs_to :category
   belongs_to :status
   belongs_to :postage
   belongs_to :sendday
   belongs_to :area
 
-  # active_storageとのアソシエーション
-  # （items・active_storage_blobsテーブルを関連付け）
   has_one_attached :image
 
   with_options presence: true do
-    validates :user_id
     validates :image
     validates :item_name
     validates :explain
@@ -25,11 +20,9 @@ class Item < ApplicationRecord
     validates :postage_id
     validates :area_id
     validates :send_day_id
-    # 300円以上かつ9,999,999円以下で、半角数字でないと入力不可
     validates :price, numericality: { greater_than_or_equal_to: 300, less_than_or_equal_to: 9_999_999 }
   end
 
-  # ジャンルの選択が「--」の時は保存不可
   with_options numericality: { other_than: 0 } do
     validates :category_id
     validates :status_id
