+++ app/models/postage.rb
@@ -1,6 +1,6 @@
 class Postage < ActiveHash::Base
   self.data = [
-    { id: 0, name: '--' },
+    { id: 0, name: '---' },
     { id: 1, name: '着払い（購入者負担）' },
     { id: 2, name: '送料払い（出品者負担）' }
   ]
